# api/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogEntry": {
        "description": "One card on the public catalog shelf.\n\nDeliberately narrower than ``ExpertDetail``: no id, no owner, no error, no\nbuild internals. This is the anonymous-readable projection.",
        "properties": {
          "avg_quality": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Quality"
          },
          "blurb": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blurb"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "chunk_count": {
            "default": 0,
            "title": "Chunk Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "graph_expanded": {
            "default": false,
            "title": "Graph Expanded",
            "type": "boolean"
          },
          "is_featured": {
            "default": false,
            "title": "Is Featured",
            "type": "boolean"
          },
          "key_concepts": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Key Concepts",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "node_count": {
            "default": 0,
            "title": "Node Count",
            "type": "integer"
          },
          "persona_bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Bio"
          },
          "persona_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Name"
          },
          "picture": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExpertPictureOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "readiness": {
            "default": "pending",
            "title": "Readiness",
            "type": "string"
          },
          "source_count": {
            "default": 0,
            "title": "Source Count",
            "type": "integer"
          },
          "source_type_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "default": {},
            "title": "Source Type Counts",
            "type": "object"
          },
          "tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "name",
          "topic",
          "tier",
          "created_at"
        ],
        "title": "CatalogEntry",
        "type": "object"
      },
...
      "ExpertAvatar": {
        "description": "A rendering recipe, not an image. See peritus.experts.avatar for why.\n\n``seed`` is optional; the client derives it when absent. ``hue`` is kept for\nolder clients and is always null \u2014 there are no per-expert colours, and a\nvalue sent here is discarded rather than refused.",
        "properties": {
          "hue": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hue"
          },
          "seed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "style": {
            "title": "Style",
            "type": "string"
          }
        },
        "required": [
          "style"
        ],
        "title": "ExpertAvatar",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeResponse": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_admin": {
            "title": "Is Admin",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "is_admin"
        ],
        "title": "MeResponse",
        "type": "object"
      },
      "OAuthExchangeRequest": {
        "properties": {
          "auth_code": {
            "minLength": 1,
            "title": "Auth Code",
            "type": "string"
          },
          "code_verifier": {
            "maxLength": 128,
            "minLength": 43,
            "title": "Code Verifier",
            "type": "string"
          }
        },
        "required": [
          "auth_code",
          "code_verifier"
        ],
        "title": "OAuthExchangeRequest",
        "type": "object"
      },
      "OtpRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "OtpRequest",
        "type": "object"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "Session": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/SessionUser"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "user"
        ],
        "title": "Session",
        "type": "object"
      },
      "SessionUser": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "SessionUser",
        "type": "object"
      },
...
      "SharedExpertOut": {
        "description": "What anyone holding a live share link sees, signed in or not.\n\nThe link preview and the share page render from this. It is the catalog\ncard's shape minus curation, plus the avatar recipe: no slug, no owner, no\nerror, no build internals \u2014 nothing the link's holder could not see after\nopening it, and nothing about who shared it.",
        "properties": {
          "avatar": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExpertAvatar"
              },
              {
                "type": "null"
              }
            ]
          },
          "avg_quality": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Quality"
          },
          "build_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Build Active"
          },
          "chunk_count": {
            "default": 0,
            "title": "Chunk Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "graph_expanded": {
            "default": false,
            "title": "Graph Expanded",
            "type": "boolean"
          },
          "key_concepts": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Key Concepts",
            "type": "array"
          },
          "node_count": {
            "default": 0,
            "title": "Node Count",
            "type": "integer"
          },
          "persona_bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Bio"
          },
          "persona_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Name"
          },
          "picture": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExpertPictureOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "readiness": {
            "default": "pending",
            "title": "Readiness",
            "type": "string"
          },
          "source_count": {
            "default": 0,
            "title": "Source Count",
            "type": "integer"
          },
          "source_type_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "default": {},
            "title": "Source Type Counts",
            "type": "object"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "topic",
          "tier",
          "created_at"
        ],
        "title": "SharedExpertOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifyRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "email",
          "token"
        ],
        "title": "VerifyRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Peritus API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/logout": {
      "post": {
        "description": "Revoke the caller's Supabase session so the refresh token can't be reused.\n\nBest-effort: clients also drop their local session. A missing/expired token is\ntreated as already-logged-out (204) rather than an error.",
        "operationId": "logout_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "me_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/oauth/authorize": {
      "get": {
        "description": "Return the GoTrue authorize URL for the web app to redirect the browser to.\n\nThe web front-end holds the PKCE verifier; we only see its challenge. GoTrue\nenforces its own redirect_to allowlist, so an attacker-supplied redirect_to\ncan't leak the auth code to a foreign origin.",
        "operationId": "oauth_authorize_auth_oauth_authorize_get",
        "parameters": [
          {
            "in": "query",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "code_challenge",
            "required": true,
            "schema": {
              "title": "Code Challenge",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "redirect_to",
            "required": true,
            "schema": {
              "title": "Redirect To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Oauth Authorize Auth Oauth Authorize Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Oauth Authorize",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/oauth/exchange": {
      "post": {
        "description": "Trade an OAuth PKCE code for a session after the provider redirect.",
        "operationId": "oauth_exchange_auth_oauth_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Oauth Exchange",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/otp": {
      "post": {
        "operationId": "send_otp_auth_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Otp",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Exchange a refresh token for a new session.\n\nTransport failures are reported as 503, never 500. The distinction is the\nclient's only way to tell \"this refresh token is dead, log out\" from \"we\ncould not reach the auth server, try again\" \u2014 and clients act on it by\nerasing the stored session. A DNS blip here once surfaced as an unhandled\n500 and logged a user out mid-build, discarding a refresh token that was\nstill perfectly valid.",
        "operationId": "refresh_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/verify": {
      "post": {
        "operationId": "verify_otp_auth_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Otp",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/catalog": {
      "get": {
        "description": "The curated public shelf. Readable without a session.",
        "operationId": "list_catalog_catalog_get",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "featured",
            "required": false,
            "schema": {
              "default": false,
              "title": "Featured",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogEntry"
                  },
                  "title": "Response List Catalog Catalog Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Catalog",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/catalog/{slug}": {
      "get": {
        "description": "One catalog card. Public experts only \u2014 sharing a private expert is a\ntoken link (``GET /share/{token}``), never its slug.",
        "operationId": "get_catalog_expert_catalog__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Catalog Expert",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/share/{token}": {
      "get": {
        "description": "The share card. Readable without a session, so a link can unfurl.",
        "operationId": "get_shared_expert_share__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedExpertOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Shared Expert",
        "tags": [
          "sharing"
        ]
      }
    },
...
  }
}